✨ feat: Refactor Artifacts components for improved structure and performance; remove unused files and optimize styles
